chore: moves getters to views
